test: add preconnect tests and move error tests
