+++ test/development/acceptance-app/fixtures/rsc-build-errors/app/metadata/mutate/page.js
@@ -1,0 +1,11 @@
+export default function page(props) {
+  return <p>mutate</p>
+}
+
+export async function generateMetadata(props, parent) {
+  const parentMetadata = await parent
+
+  return {
+    ...parentMetadata,
+  }
+}
