+++ test/e2e/app-dir/metadata/app/basic/client.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import ReactDOM from 'react-dom'
+
+// NOTE: atm preload/prefetch/prefetchDNS are only supported in the SSR and client, not in RSC yet
+export default function Client() {
+  // @ts-ignore
+  ReactDOM.preload('/preload-url', { as: 'script' })
+  // @ts-ignore
+  ReactDOM.preconnect('/preconnect-url', { crossOrigin: 'use-credentials' })
+  // @ts-ignore
+  ReactDOM.prefetchDNS('/dns-prefetch-url')
+  return null
+}
